tobefair_backend_requirements.txt
bs4==0.0.2
extruct==0.16.0
jmespath==1.0.1
fastapi==0.121.0
python-mimeparse==1.6.0
tika==2.6.0
types-jmespath==1.0.2.20240106
uvicorn==0.29.0